Update README.md
Since tweets are essentially a source used by people to express opinions, it surely contains one of the most vital characteristics of language; sentiments. And thus, with this project, I try to analyse the positive and negative sentiments among a large pool of tweets.

Using Natural Language Processing, sentiments in the text can be analysed. In this, I'll be using Naive Bayes Classification to classify the tweet as 'Positive' or 'Negative'. But before performing this classification, I'll be exploring the dataset and also plotting a WordCloud.